backend/requirements.txt
Django==5.0.3
djangorestframework==3.16.0
django-cors-headers==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0 